+++ src/components/MUIVideoTile.js
@@ -3,6 +3,7 @@ import React from 'react';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
+import './VideoTile.css'
 
 
 export const MUIVideoTile = (props) => {
